+++ tests/test_visitors/test_ast/test_builtins/test_numbers/test_magic_numbers.py
@@ -29,6 +29,18 @@ dict_definition_value = '{{"first": {0}}}'
 set_definition = '{{"first", {0}, "other"}}'
 tuple_definition = '({0}, )'
 
+literal_type_hint = 'code: Literal[{0}] = {0}'
+literal_type_hint_with_typing_module = 'code: typing.Literal[{0}] = {0}'
+literal_type_hint_with_typing_exts_module = (
+    'code: typing_extensions.Literal[{0}] = {0}'
+)
+literal_type_hint_without_assign = 'code: Literal[{0}]'
+literal_type_hint_typing_without_assign = 'code: typing.Literal[{0}]'
+literal_type_hint_typing_ext_without_assign = (
+    'code: typing_extensions.Literal[{0}]'
+)
+
+
 # Wrong usages:
 
 assignment_binop = 'final = {0} + 1'
@@ -68,6 +80,13 @@ some_dict = {{11: 12, 13: 14}}
 some_dict[{0}]
 """
 
+not_literal_type_hint = """
+foo: Bar[{0}] = {0}
+"""
+
+literal_type_with_not_typing_module = 'foo: bar.Literal[{0}] = {0}'
+literal_type_not_typing_module_no_assign = 'foo: bar.Literal[{0}]'
+
 
 @pytest.mark.parametrize(
     'code',
@@ -83,6 +102,12 @@ some_dict[{0}]
         dict_definition_value,
         set_definition,
         tuple_definition,
+        literal_type_hint,
+        literal_type_hint_with_typing_module,
+        literal_type_hint_with_typing_exts_module,
+        literal_type_hint_without_assign,
+        literal_type_hint_typing_without_assign,
+        literal_type_hint_typing_ext_without_assign,
     ],
 )
 @pytest.mark.parametrize(
@@ -145,6 +170,8 @@ def test_magic_number(
         inside_method,
         list_index,
         dict_key,
+        not_literal_type_hint,
+        literal_type_with_not_typing_module,
     ],
 )
 @pytest.mark.parametrize(
@@ -194,6 +221,9 @@ def test_magic_number_whitelist(
         inside_method,
         list_index,
         dict_key,
+        not_literal_type_hint,
+        literal_type_with_not_typing_module,
+        literal_type_not_typing_module_no_assign,
     ],
 )
 @pytest.mark.parametrize(
@@ -244,6 +274,9 @@ def test_magic_number_warning(
         inside_method,
         list_index,
         dict_key,
+        not_literal_type_hint,
+        literal_type_with_not_typing_module,
+        literal_type_not_typing_module_no_assign,
     ],
 )
 @pytest.mark.parametrize(